% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dates.R
\name{basic_date}
\alias{basic_date}
\title{Create a basic date from date time object}
\usage{
basic_date(datetimes)
}
\arguments{
\item{datetimes}{Object of class \code{"POSIXct/POSIXt"}, can have multiple elements.}
}
\value{
Character vector of length equal to that of \code{datetimes}, but converted to format \code{"M/D/YYYY"}.
}
\description{
Convert a \code{"POSIXct/POSIXt"} object to a character
class that is easier to read for humans.
}
