% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-misc.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a character string}
\usage{
capitalize(x)
}
\arguments{
\item{x}{Character; a vector of character strings for which the
first character should be converted to uppercase. If not
a character, will be coerced to one.}
}
\description{
Applies \code{\link[base:chartr]{base::toupper()}} to the first character in each
element of a character vector
}
