% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dates.R
\docType{data}
\name{dates}
\alias{dates}
\title{Convert date to 'days past May 31'}
\format{
An object of class \code{Date} of length 4.
}
\usage{
dates
}
\arguments{
\item{dates}{Datetime; can be of length > 1}
}
\description{
Standardizes a \code{datetime} object to be the number of days
past May 31 of that year
}
\keyword{datasets}
