% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-axes.R
\name{draw_axis_line}
\alias{draw_axis_line}
\title{Draw an axis line with no ticks or labels}
\usage{
draw_axis_line(side, col = par("col.axis"))
}
\arguments{
\item{side}{Numeric; the side of the plot to draw the axis on;
1 = bottom, 2 = left, 3 = top, 4 = right}

\item{col}{Character; color to use for the axis line, tick marks, and tick mark labels (where applicable).
Defaults to \code{par("col.axis")}.}
}
\description{
Draw an axis line with no ticks or labels
}
