% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-axes.R
\name{draw_day_axis}
\alias{draw_day_axis}
\title{Draw an axis showing dates}
\usage{
draw_day_axis(fday, lday, by, side = 1, col = par("col.axis"), ...)
}
\arguments{
\item{fday}{Numeric; the first day on the axis to draw (1 = June 1; see \code{\link[=to_days_past_may31]{to_days_past_may31()}})}

\item{lday}{Numeric; the last day (30 = June 30; see \code{\link[=to_days_past_may31]{to_days_past_may31()}})}

\item{by}{Numeric; interval to draw axis ticks at}

\item{side}{Numeric; the side of the plot to draw the axis on;
1 = bottom, 2 = left, 3 = top, 4 = right}

\item{col}{Character; color to use for the axis line, tick marks, and tick mark labels (where applicable).
Defaults to \code{par("col.axis")}.}

\item{...}{Optional arguments to be passed to \code{\link[graphics:axis]{graphics::axis()}}}
}
\description{
Uses date labels, but for plots with user coordinates
based on the number of days past May 31st.
}
