% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-axes.R
\name{draw_percent_axis}
\alias{draw_percent_axis}
\title{Draw an axis showing percentages}
\usage{
draw_percent_axis(side, at = NULL, col = par("col.axis"), ...)
}
\arguments{
\item{side}{Numeric; the side of the plot to draw the axis on;
1 = bottom, 2 = left, 3 = top, 4 = right}

\item{at}{Numeric; optional vector to place tick marks and tick mark labels on the proportional scale.
Defaults to \code{NULL}, in which case \code{\link[grDevices:axisTicks]{grDevices::axisTicks()}} is used to automatically select tick mark locations.}

\item{col}{Character; color to use for the axis line, tick marks, and tick mark labels (where applicable).
Defaults to \code{par("col.axis")}.}

\item{...}{Optional arguments to be passed to \code{\link[graphics:axis]{graphics::axis()}}}
}
\description{
Uses percentage labels, but for plots with user coordinates
based on the proportional scale.
}
