% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-axes.R
\name{draw_regular_axis}
\alias{draw_regular_axis}
\title{Draw a regular axis}
\usage{
draw_regular_axis(side, col = par("col.axis"), ...)
}
\arguments{
\item{side}{Numeric; the side of the plot to draw the axis on;
1 = bottom, 2 = left, 3 = top, 4 = right}

\item{col}{Character; color to use for the axis line, tick marks, and tick mark labels (where applicable).
Defaults to \code{par("col.axis")}.}

\item{...}{Optional arguments to be passed to \code{\link[graphics:axis]{graphics::axis()}}}
}
\description{
Behaves just like \code{\link[graphics:axis]{graphics::axis()}}
except handles colors just like
\code{\link[=draw_day_axis]{draw_day_axis()}}, \code{\link[=draw_percent_axis]{draw_percent_axis()}}, and \code{\link[=draw_yn_axis]{draw_yn_axis()}}.
}
