% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-axes.R
\name{draw_yn_axis}
\alias{draw_yn_axis}
\title{Draw an axis showing yes/no categories}
\usage{
draw_yn_axis(side, col = par("col.axis"), ...)
}
\arguments{
\item{side}{Numeric; the side of the plot to draw the axis on;
1 = bottom, 2 = left, 3 = top, 4 = right}

\item{col}{Character; color to use for the axis line, tick marks, and tick mark labels (where applicable).
Defaults to \code{par("col.axis")}.}

\item{...}{Optional arguments to be passed to \code{\link[graphics:axis]{graphics::axis()}}}
}
\description{
Uses character labels, but for plots with user coordinates based on
0 = No, 1 = Yes.
}
