% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dates.R
\name{file_date}
\alias{file_date}
\title{Create a date for use in file names}
\usage{
file_date(x)
}
\arguments{
\item{x}{and object of class \code{datetime}}
}
\description{
Converts a \code{datetime} object to a
character class with format \code{YYYY_MM_DD}. This format
is useful in constructing file names.
}
