% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-misc.R
\name{get_errors}
\alias{get_errors}
\title{Obtain Prediction Errors and Summaries}
\usage{
get_errors(yhat, yobs)
}
\arguments{
\item{yhat}{Numeric vector of predicted values.}

\item{yobs}{Numeric vector of observed values.}
}
\value{
A \code{\link[base:list]{list}} object with elements \verb{$error} (\code{yhat - yobs}), \verb{$p_error} (\code{(yhat - yobs)/yobs}), and \verb{$summary}.
\verb{$summary} is a named vector with elements:
\itemize{
\item \code{RHO} -- Pearson correlation coefficient between \code{yhat} and \code{yobs}.
\item \code{RMSE} -- Root mean squared error: \code{sqrt(mean(error^2))}.
\item \code{ME} -- Mean error: \code{mean(error)}.
\item \code{MAE} -- Mean absolute error: \code{mean(abs(error))}.
\item \code{MPE} -- Mean proportional error: \code{mean(p_error)}.
\item \code{MAPE} -- Mean absolution proportional error: \code{mean(abs(p_error))}.
}
}
\description{
Obtain Prediction Errors and Summaries
}
\note{
If \code{yhat} and \code{yobs} are of unequal lengths, the function will fail with a useful error message.
}
