% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-misc.R
\name{get_Ewind}
\alias{get_Ewind}
\title{Obtain Easterly Wind Speed Vector}
\usage{
get_Ewind(speed, angle, digits = 1)
}
\arguments{
\item{speed}{Numeric; wind speed associated with the angle}

\item{angle}{Numeric; wind directional angle associated with the speed.}

\item{digits}{Numeric; how many decimal places to round to? Passed to \code{\link[base:Round]{base::round()}} and defaults to 1}
}
\value{
A numeric value representing the vector of easterly wind speed.
Positive values correspond to easterly winds, negative values to westerly winds.
The absolute magnitude of the number corresponds to the speed in the east or west direction.
}
\description{
Decomposes a directional wind speed vector into its easterly speed
}
\details{
The \code{angle} represents degrees from exactly northerly wind:
\itemize{
\item \code{angle = 0}: exactly northerly wind (i.e., from north to south, blowing in your face standing north)
\item \code{angle = 45}: exactly northeasterly wind
\item \code{angle = 180}: exactly southerly wind
}
}
