% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-misc.R
\name{get_Nwind}
\alias{get_Nwind}
\title{Obtain northerly wind speed vector}
\source{
G. Decossas provided this formula based on some previous work.
}
\usage{
get_Nwind(speed, angle, digits = 1)
}
\arguments{
\item{speed}{Numeric; wind speed associated with \code{angle}.}

\item{angle}{Numeric; wind directional angle associated with \code{speed}.}

\item{digits}{Numeric; how many decimal places to round to? Passed to \code{\link[base:Round]{base::round()}} and defaults to 1.}
}
\value{
A numeric value representing the vector of northerly wind speed.
Positive values correspond to northerly winds, negative values to southerly winds.
The absolute magnitude of the number corresponds to the speed in the north or south direction.
}
\description{
Decomposes a directional wind speed vector into its northerly speed
}
\details{
The \code{angle} represents degrees from exactly northerly wind:
\itemize{
\item \code{angle = 0}: exactly northerly wind (i.e., from north to south, blowing in your face standing north)
\item \code{angle = 45}: exactly northeasterly wind
\item \code{angle = 180}: exactly southerly wind
}
}
\note{
The calculation performed is:
\deqn{\text{speed} \cdot \text{cos}(\pi\cdot\frac{\text{angle}}{180})}
}
