% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-misc.R
\name{get_period}
\alias{get_period}
\title{Obtain the "Period" of the Season the Record Falls In}
\usage{
get_period(x)
}
\arguments{
\item{x}{Either a numeric value representing days past May 31 that year
or a date object, which will be coerced to days past May 31 prior to the calculation.}
}
\value{
The period number corresponding to the input date(s) supplied to \code{x}:
\itemize{
\item \code{1}: June 12 - June 19; first week of drift fishing allowed.
\item \code{2}: June 20 - June 30; remainder of June.
\item \code{3}: July 1 - July 30; any date in July.
\item \code{NA}: if the date does not fall in any of these periods.
}
}
\description{
Quickly assigns a time period (three-levels) of the season
based on the date
}
