% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-misc.R
\name{make_period_labels}
\alias{make_period_labels}
\title{Make period labels}
\usage{
make_period_labels(last_day = 61)
}
\arguments{
\item{last_day}{Numeric; last day to use as the end of the July period. Must be >= 31; defaults to 61 (July 31).}
}
\value{
Character vector with the date ranges for each period, as defined by \code{\link[=get_period]{get_period()}}.
}
\description{
Quickly obtains range of dates included in each period.
}
