% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dates.R
\name{short_datetime}
\alias{short_datetime}
\title{Format a datetime object to be shorter}
\usage{
short_datetime(datetimes, include_date = FALSE)
}
\arguments{
\item{datetimes}{Object of class \code{"POSIXct/POSIXt"}, can have multiple elements.}

\item{include_dates}{Logical; whether the date should be included, as well as the time.}
}
\value{
Character vector of length equal to that of \code{datetimes}.
If \code{include_dates = FALSE}, only the time will be returned (12-hr clock).
If \code{include_dates = TRUE}, the time and the date will be returned (\code{"M/D"} format).
}
\description{
Convert a \code{"POSIXct/POSIXt"} object to a character
class that is easier to read for humans.
}
