% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-TeX.R
\name{tinyCI}
\alias{tinyCI}
\title{Make the CI part of a summary smaller text than the mean}
\usage{
tinyCI(x, linebreak = TRUE)
}
\arguments{
\item{x}{Character vector; the output of \code{\link[KuskoHarvEst:report]{KuskoHarvEst::report()}}, or any character string with the format \code{"est (lwrCI-uprCI)"}.}

\item{linebreak}{Logical; should a line break be inserted between the mean portion and the CI portion?}
}
\value{
Character vector with each element correctly formatted with LaTeX code.
}
\description{
Adds the appropriate LaTeX code to a summary produced by
\code{\link[KuskoHarvEst:report]{KuskoHarvEst::report()}} to make the CI portion smaller than the mean portion.
This is a nice touch for tabular output.
}
