% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dates.R
\name{basic_date}
\alias{basic_date}
\title{Create a basic date from date time object}
\usage{
basic_date(datetime)
}
\arguments{
\item{datetime}{Object of class \code{datetime}}
}
\description{
Convert a \code{datetime} object to a character
class that is easier to read for humans.
}
