% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dates.R
\name{file_date}
\alias{file_date}
\title{Create a date for use in file names}
\usage{
file_date(datetimes)
}
\arguments{
\item{datetimes}{Object of class \code{"POSIXct/POSIXt"}, can have multiple elements.}
}
\value{
Character vector of length equal to that of \code{datetimes}, but converted to format \code{"YYYY_MM_DD"}.
}
\description{
Converts a \code{"POSIXct/POSIXt"} object to a
character class useful in constructing file names.
}
