% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dates.R
\name{from_days_past_may31}
\alias{from_days_past_may31}
\title{Convert 'days past May 31' to a date}
\usage{
from_days_past_may31(days, year = "1900")
}
\arguments{
\item{year}{Numeric or character; the year of the output \code{datetime} object. Can be of length > 1, but if so, should be the same length as \code{day}}

\item{day}{Numeric; number of days past May 31 to convert to a \code{datetime} object. Can be of length > 1}
}
\description{
Unstandardizes a number of days past May 31
to be a \code{datetime} object
}
