% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-misc.R
\name{get_errors}
\alias{get_errors}
\title{Obtain prediction errors and summaries}
\usage{
get_errors(yhat, yobs, FUN = median)
}
\arguments{
\item{yhat}{Numeric; vector of predicted values.}

\item{yobs}{Numeric; vector of observed values.}

\item{FUN}{Function used to summarize the central tendency of errors;
defaults to \code{\link[stats:median]{stats::median()}} to minimize influence of rare large errors but \code{\link[base:mean]{base::mean()}} is another good option.}
}
\value{
A \code{\link[base:list]{list}} object with elements \verb{$error} (\code{yhat - yobs}), \verb{$p_error} (\code{(yhat - yobs)/yobs}), and \verb{$summary}.
\verb{$summary} is a named vector with elements:
\itemize{
\item \code{RHO} -- Pearson correlation coefficient between \code{yhat} and \code{yobs}.
\item \code{RMSE} -- Root \code{FUN} squared error: \code{sqrt(FUN(error^2))}.
\item \code{ME} -- \code{FUN} error: \code{FUN(error)}.
\item \code{MAE} -- \code{FUN} absolute error: \code{FUN(abs(error))}.
\item \code{MPE} -- \code{FUN} proportional error: \code{FUN(p_error)}.
\item \code{MAPE} -- \code{FUN} absolution proportional error: \code{FUN(abs(p_error))}.
}
}
\description{
Given predicted and observed vectors,
calculates errors and a range of summary statistics.
}
\note{
If \code{yhat} and \code{yobs} are of unequal lengths, the function will fail with a useful error message.
}
