% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-TeX.R
\name{kable_replace}
\alias{kable_replace}
\title{Replace Text in a kable Object}
\usage{
kable_replace(kable_input, pattern, replacement)
}
\arguments{
\item{kable_input}{The output of a Output of a \code{knitr::kable() |> kableExtra::fn()} chain.}

\item{pattern}{Pattern to look for -- interpreted as a regular expression. See \code{\link[stringr:str_replace]{stringr::str_replace()}}.}

\item{replacement}{Replacement value.}
}
\description{
Sometimes something needs to be altered
in the TeX code after the kable object is built.
}
