% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-misc.R
\name{percentize}
\alias{percentize}
\title{Convert a proportion to a percent}
\usage{
percentize(x, escape = FALSE, digits = 0)
}
\arguments{
\item{x}{Numeric; vector containing values on the proportional scale to be converted to a percentage value}

\item{escape}{Logical; should the percent symbol be escaped? Aids in placing this output into LaTeX tables}

\item{digits}{Numeric; supplied to \code{\link[base:Round]{base::round()}}. Defaults to \code{0}.}
}
\value{
Character vector storing percentage values. If a non-zero value would be rounded to zero, \code{"<1\%"} is returned instead.
}
\description{
Convert a proportion to a percent
}
