% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-TeX.R
\name{tinyCI}
\alias{tinyCI}
\title{Make the CI part of a summary smaller text than the mean}
\usage{
tinyCI(x, linebreak = TRUE)
}
\arguments{
\item{x}{Character; the output of \code{\link[=report]{report()}}}

\item{linebreak}{Logical; should a linebreak be inserted between the mean portion and the CI portion?}
}
\description{
Adds the appropriate LaTeX code to a summary produced by
\code{\link[=report]{report()}} to make the CI portion smaller than the mean portion.
This is a nice touch for tabular output.
}
