% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-dates.R
\name{to_days_past_may31}
\alias{to_days_past_may31}
\title{Convert date to 'days past May 31'}
\usage{
to_days_past_may31(dates)
}
\arguments{
\item{dates}{Datetime; can be of length > 1}
}
\description{
Standardizes a \code{datetime} object to be the number of days
past May 31 of that year
}
